backend/requirements.txt
pymysql
fastapi==0.116.1
uvicorn[standard]
celery
SQLAlchemy==2.0.43
python-dotenv==1.1.1
redis==5.2.0
cryptography
pandas